+++ lib/src/ever_value_base.dart
@@ -1,3 +1,7 @@
+import 'ever_cached_state.dart';
+
 abstract class IEverValue<T> {
   T get value;
+
+  EverCachedState<T> get state;
 }