+++ CHANGELOG.md
@@ -22,3 +22,12 @@
 ## 0.0.6
 
 - Fix the issue with value getter
+
+## 0.0.7
+
+- Introduce EverCachedValue<T> class to wrap the actual to prevent cases which can occur if compute method returned null
+
+## 0.0.8
+
+- Fix the issue with EverCachedValue<T> class
+- Refactors